Sync internal 625abf99 (#35)
* [compiler/cmake] Added asan support, and updated MSVC build 
* [compiler/gpu] Added gpu arch specifier, and improved GPU codegen
* [compiler] Supported more broadcast fusion, and improved elementwise codegen
* [compiler/transformation] Supported more loop tag, and improved memory planning
* [compiler/utils] Improved cloning, folding, and hoist
* [compiler/CAT] Improved AIT supports in CAT
* [compiler/mhlo] Added canonicalizers
* [compiler] Fixed bugs
* [compiler/E2E] Made default pipelines from affine to scf
* [compiler] Bumped version to 1.1.0
* [frontend/onnx] Fixed readme
* [frontend/tf] Improved TF frontend attributes
* [frontend/torch] Bumped torch-mlir version
* [runtime] Bumped version to 1.1.0
* [runtime] Added TF attribute support

+++ compiler/lib/Transforms/LoopTag.cpp
@@ -30,7 +30,7 @@ namespace {
 
 struct LoopTagPass : public LoopTagBase<LoopTagPass> {
   LoopTagPass(const std::string &anchor, const std::string &attach,
-              unsigned depth, const std::string &loopType)
+              int64_t depth, const std::string &loopType)
       : LoopTagBase<LoopTagPass>() {
     this->anchorAttr = anchor;
     this->attachAttr = attach;
@@ -77,7 +77,7 @@ struct LoopTagPass : public LoopTagBase<LoopTagPass> {
 
 std::unique_ptr<OperationPass<func::FuncOp>>
 mlir::createLoopTagPass(llvm::StringRef anchorTag, const std::string &attachTag,
-                        unsigned depth, const std::string &loopType) {
+                        int64_t depth, const std::string &loopType) {
   return std::make_unique<LoopTagPass>(anchorTag.str(), attachTag, depth,
                                        loopType);
 }
